requirements.txt
# Core Reddit scraping and data processing
praw>=7.7.1
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Language detection and translation
langdetect>=1.0.9
googletrans==4.0.0rc1
deep-translator>=1.11.4

# Network analysis
networkx>=3.1
matplotlib>=3.7.0
plotly>=5.15.0
dash>=2.12.0

# Machine learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
transformers>=4.30.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web interface
gradio>=3.35.0
streamlit>=1.25.0

# Data visualization
seaborn>=0.12.0
wordcloud>=1.9.0

# Text processing
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
loguru>=0.7.0


# Vector database support
pgvector>=0.2.0

# Embeddings and semantic analysis
sentence-transformers>=2.2.0
openai>=1.0.0  # For embeddings API if needed

# Database migration support
alembic>=1.12.0
